Move Fieldset into own file
